chore(docs): folders and items

+++ src/folders/folders.controller.ts
@@ -13,11 +13,20 @@ import {
   Put,
   Logger,
   Query,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { FoldersService } from './folders.service';
 import { CreateFolderDto } from './dto/create-folder.dto';
 import { UpdateFolderDto } from './dto/update-folder.dto';
-import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import {
+  ApiBadRequestResponse,
+  ApiBearerAuth,
+  ApiNotFoundResponse,
+  ApiOperation,
+  ApiQuery,
+  ApiTags,
+  ApiUnauthorizedResponse,
+} from '@nestjs/swagger';
 import { OrganizationGuard } from '../auth/organization.guard';
 import { AuthenticatedRequest } from '../shared/types/authentication.types';
 import { Folder } from './schemas/folder.schema';
@@ -27,6 +36,8 @@ import { BlocksService } from '../blocks/blocks.service';
 import { Block } from '../blocks/schemas/block.schema';
 import { FolderItem } from './schemas/folder-item.schema';
 import { FolderItemType } from './entities/folder-item-type';
+import { Action } from '../rbac/permissions/entities/action.entity';
+import { ScopedSubject } from '../rbac/casl/casl.utils';
 
 @ApiTags('Folders')
 @ApiBearerAuth()
@@ -42,14 +53,21 @@ export class FoldersController {
   ) {}
 
   @Post()
-
+  @ApiOperation({ summary: 'Create new folder' })
+  @ApiUnauthorizedResponse({
+    description: 'Cannot create folder, requires "Create Folder" permission',
+  })
+  @ApiNotFoundResponse({ description: 'Parent folder not found' })
   async create(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Body() createFolderDto: CreateFolderDto,
-  ) {
-
-    if (req.ability.cannot(Action.Create, ScopedSubject(Folder, organizationId))) {
+  ): Promise<Folder> {
+    if (
+      req.ability.cannot(Action.Create, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot create folder');
+    }
 
     if (createFolderDto.parentFolder) {
       const parentFolder = await this.foldersService.findOne(
@@ -66,29 +84,60 @@ export class FoldersController {
   }
 
   @Get()
+  @ApiOperation({ summary: 'Get all folders' })
+  @ApiUnauthorizedResponse({
+    description: 'Cannot read folders, requires "Read Folder" permission',
+  })
   findAll(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
   ): Promise<Folder[]> {
+    if (
+      req.ability.cannot(Action.Read, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot read folders');
+    }
+
     return this.foldersService.findAllOfOrganization(organizationId);
   }
 
   @Get(':folderId')
+  @ApiOperation({ summary: 'Get folder by ID' })
+  @ApiUnauthorizedResponse({
+    description: 'Cannot read folder, requires "Read Folder" permission',
+  })
   findOne(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('folderId') folderId: string,
-  ) {
+  ): Promise<Folder> {
+    if (
+      req.ability.cannot(Action.Read, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot read folder');
+    }
+
     return this.foldersService.findOne(organizationId, folderId);
   }
 
   @Patch(':folderId')
+  @ApiOperation({ summary: 'Update folder by ID' })
+  @ApiUnauthorizedResponse({
+    description: 'Cannot update folder, requires "Update Folder" permission',
+  })
+  @ApiNotFoundResponse({ description: 'Parent folder not found' })
   async update(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('folderId') folderId: string,
     @Body() updateFolderDto: UpdateFolderDto,
-  ) {
+  ): Promise<Folder> {
+    if (
+      req.ability.cannot(Action.Update, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot update folder');
+    }
+
     if (updateFolderDto.parentFolder) {
       const parentFolder = await this.foldersService.findOne(
         organizationId,
@@ -108,17 +157,31 @@ export class FoldersController {
   }
 
   @Delete(':folderId')
+  @ApiOperation({ summary: 'Delete folder by ID, and sub-folders recursively' })
+  @ApiUnauthorizedResponse({
+    description: 'Cannot delete folder, requires "Delete Folder" permission',
+  })
   remove(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('folderId') folderId: string,
   ) {
+    if (
+      req.ability.cannot(Action.Delete, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot delete folder');
+    }
     return this.foldersService.remove(organizationId, folderId, true);
   }
 
   // Folder Items
 
   @Put(':folderId/items/:itemId')
+  @ApiOperation({ summary: 'Add item to folder' })
+  @ApiUnauthorizedResponse({
+    description:
+      'Cannot add item to folder, requires "Update Folder" permission',
+  })
   async addFolderItem(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
@@ -126,11 +189,17 @@ export class FoldersController {
     @Param('itemId') itemId: string,
     @Body() itemDto: FolderItemDto,
   ): Promise<FolderItem> {
+    if (
+      req.ability.cannot(Action.Update, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot add item to folder');
+    }
+
     itemDto.folder = folderId;
     itemDto.organization = organizationId;
     itemDto.item = itemId;
 
-    const item = await this.folderItemsService.findOne(
+    const item: FolderItem = await this.folderItemsService.findOne(
       organizationId,
       folderId,
       itemId,
@@ -141,20 +210,42 @@ export class FoldersController {
   }
 
   @Get(':folderId/items')
+  @ApiOperation({ summary: 'Get items of folder' })
+  @ApiQuery({
+    name: 'type',
+    enum: FolderItemType,
+    required: false,
+    description: 'Filter items by type',
+  })
+  @ApiBadRequestResponse({
+    description: 'Item type not supported, check FolderItemType',
+  })
+  @ApiUnauthorizedResponse({
+    description:
+      'Cannot get items of folder, requires "Read Folder" permission',
+  })
   async getFolderItems(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('folderId') folderId: string,
     @Query('type') type?: FolderItemType,
   ): Promise<Array<Block>> {
-    let items = await this.folderItemsService.findAllOfFolder(
+    if (
+      req.ability.cannot(Action.Read, ScopedSubject(Folder, organizationId)) ||
+      req.ability.cannot(Action.Read, ScopedSubject(Block, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot get items of folder');
+    }
+
+    let items: FolderItem[] = await this.folderItemsService.findAllOfFolder(
       organizationId,
       folderId,
     );
 
     if (type) items = items.filter((i) => i.type === type);
 
-    const itemsPopulated = [];
+    // Maybe we should create an abstract type to handle plurality of items
+    const itemsPopulated: Block[] = [];
 
     for (const item of items) {
       switch (item.type) {
@@ -176,12 +267,26 @@ export class FoldersController {
   }
 
   @Delete(':folderId/items/:itemId')
+  @ApiOperation({
+    summary: 'Remove item from folder by ID, does not delete item',
+  })
+  @ApiUnauthorizedResponse({
+    description:
+      'Cannot remove item from folder, requires "Update Folder" permission',
+  })
+  @ApiNotFoundResponse({ description: 'Folder or item not found' })
   async removeFolderItem(
     @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('folderId') folderId: string,
     @Param('itemId') itemId: string,
   ) {
+    if (
+      req.ability.cannot(Action.Update, ScopedSubject(Folder, organizationId))
+    ) {
+      throw new UnauthorizedException('Cannot remove item from folder');
+    }
+
     const folder = await this.foldersService.findOne(organizationId, folderId);
 
     if (!folder) throw new NotFoundException('Folder not found');
